Rename mhd_wave_class.py to mhd_waves_.py and refactor
