+++ changelog/2206.feature.rst
@@ -1,3 +1,3 @@
-Added module `plasmapy.dispersion.analytical.mhd_wave_class`
+Added module `plasmapy.dispersion.analytical.mhd_waves_`
 with classes for storing and calculating parameters of
 magnetohydrodynamic waves.
