+++ plasmapy/dispersion/analytical/mhd_waves_.py
